Serial working with GUI

+++ stub.h
@@ -1,4 +1,6 @@
+#include <stdint.h>
 #include <string>
+
 void pinMode(int pin, int type);
 int digitalRead(int pin);
 void digitalWrite(int pin, int value);
@@ -35,4 +37,4 @@ long map(long x, long in_min, long in_max, long out_min, long out_max);
 
 void delay(int milliseconds);
 
-unsigned long micros();
+uint32_t micros();
